Make AssetManagerWithTieredStorage not abstract
